ui(PlotDetail): decompose 3000-line monolith onto plot-detail/ sub-tree
PlotDetail.tsx grew to 3099 lines covering header controls, intent
editing, substrate attachments, run-plan + dispatch dialogs, batch
dispatch, activity feed, and the interactive Chat panel. Phase 7 of
the design-system revamp (warren-2221 / pl-55a3 step 8) splits it
into focused sibling modules under src/ui/src/pages/plot-detail/:

- helpers.ts          — readString, isSdPlanAttachment, isBatchDispatchTarget
- header-controls.tsx — PlotSyncButton, PlotNameEditor, StatusTransitionControl
- intent-panel.tsx    — IntentPanel + IntentListField + IntentDraft helpers
- substrate-panel.tsx — SubstratePanel, AttachmentRow, AddAttachmentDialog
- run-plan.tsx        — RunPlanButton/Dialog, DispatchAsPlanRunButton/Dialog, ReadOnlyField
- batch-dispatch.tsx  — BatchDispatchAllButton/Dialog, MergeOutcomeBanner, RunSeedButton
- activity-feed.tsx   — ActivityFeed, EventLine, AnswerCard, PauseCountdown, ActorSlot
- interactive-panel.tsx — InteractivePanel, DispatchInteractiveDialog, FormalizeDialog

PlotDetailPage in PlotDetail.tsx now owns just the tanstack-query
fetch, the loading / error / 404 branches, and the layout shell — 124
lines. No behavioral changes; cross-module exports preserve the prior
component graph. helpers.ts breaks the substrate↔batch import cycle
that would otherwise emerge from moving isSdPlanAttachment.

src/ui-tests/plot-detail-run-plan-dialog.test.ts now sources from
plot-detail/run-plan.tsx and tolerates 'export function' decls.

Phase 7 still nominally blocked on warren-57d7/3849/36f0/5da1/e6b3,
but those phases' primitives are already on disk under components/ui/
— this decomposition uses them in place. New primitive adoption (Field
/ PageHeader / Alert sweeps) can land incrementally on top of the
smaller modules instead of fighting the monolith.
